Add a different calculation method for the polymer chain.
For each step, two random angles, theta and phi, are chosen. Using spherical
coordinates, the bext point is calculated. Contrary to the first version,
collisions are not allowed. Thus, if the distance to any other point is < 1,
a different point is randomly calculated. This assumes that monomers are
spheres with radius 1/2.