On Sun, Mar 27, 2005 at 11:00:38PM +0000, Thorsten Glaser wrote: > Thomas Dickey dixit: > > > ok - it's probably reasonably stable.  I had been working on fixes for > > the wide-character stuff > > It's already much better in uxterm.
yes - there are 2-3 fixes.  But see dev.11i (which I had overlooked putting
there til tonight).

That fixes something that I'd broken in 2.8.5 w/o noticing it (text
justification).

I only noticed it because I was looking closely at some of Debian's webpages,
and realized that I had justification compiled in but not working.  So it's
good (but bad that it was overlooked for so long).