Finally add a meta port for GNOME 2.
From DESCR:

This is a meta-package. It does not include any files by itself, but it
helps you to easily install the GNOME desktop.

If you build this from source, you need not even install the resulting
package because everything will have been installed as a build-time
dependency. For binary packages, installing the package will pull in all
the other packages as a run dependency.

--

Now that see the actual package, I don't think the binary package thing is
going to work because the dependencies are sorted alphabetically there,
while the build depends install them in order.