Carefully adapted diff:
From: Marc Matteo <marcm@lectroid.net>

The first was bug where -pthread wasn't getting placed in the
*.la files properly and so packages that were using libtool to
divine linking flags wouldn't pick up -pthread when they needed
it.

The other was a bug where the permissions on any directories
created during the relink process were such that a 'make clean'
as a normal user would fail if the relink process was run during
faking (which is usually as root).