Add mirbsdksh R21, which was developed in a temporary external CVS repo- sitory whose ChangeLog follows. mksh R21 is licenced under the MirOS li- cence, shown in "sh.h", and a two-clause UCB-style licence by Marc Espie as shown in "alloc.c".
This executable is a fair bit smaller and shorter than our /bin/ksh that
it is designed to eventually replace (as /bin/sh hardlink), with the old
/bin/ksh to completely vanish. It is still in beta testing though, and I
don't think it will compile on other operating systems.

mksh R21 is a completely new port, bringing together the OpenBSD-current
/bin/ksh, the MirOS-current /bin/ksh and the older mksh R20 (which still
was portable, ocvs-based).