Correct libtool version numbering, bump patchlevel again.
We inherited OpenBSD's (wrong) 1:2:0 version number. This gives us
libglib.1.0.2.dylib in the end. This, of course, does not work with
"glib.1.2" lib dependencies. Now we use 3:10:2, which gives
libglib.1.2.10.dylib.

We need to search for other such version number kludges in the ports tree.
x11/gtk+1 is a good candidate.

agreed tg@