Add -lSM -lICE to X libraries, they are dependencies of the Athena widgets.
Also remove AC_FUNC_MALLOC from the configure script. This function is more
of a design bug than useful. If malloc is not GNU malloc it defines malloc
as rpl_malloc, which is undefined. Wtf?

(It seems that when you use this macro, you are supposed to add a source
file with rpl_malloc, which is a wrapper around the platform's malloc).