Add support for dropping privileges in pkg_add.
Here is how it works: You have installed a MirPorts instance as user
(for example with Setup.sh -ule /path/to/prefix). Now if you install a
package under sudo, writes into the package database are done with the
privileges of the user, not with root rights. This means that there won't
be any problems with file permissions inside the package db.