The last commit aborted and fucking CVS ate my log message! Anyway:
Implement a new pkgspec match type called a multiversion match. It allows
processing dependencies such as
	foo-1.0.*,>=2.2,<2.3
Glob matches (with star) and relational matches (with relop) can be freely
matched. Relationals all have to match, while one of the globs is sufficient,
i.e. the above means foo-1.0.* OR (foo->=2.2 AND foo-<2.3). The last missing
feature from pkg(1), I hope.

Written in the train this morning, which was half an hour late ...