commit on behalf of bsiegert@ who doesn't yet have © papers set up with FSF:
Keep cvs from eating log messages on unexpected errors: keep temp file with the log
message after do_editor() finishes; at the end of import() and commit() (these are
the two that may call do_editor), a new procedure logmsg_cleanup() is called, which
either deletes the temp file or leaves it and tells the user the file name.

The idea is good. We should feed that back to upstream.