Add a port for YAML::Syck. From DESCR:
This module provides a Perl interface to the libsyck data serialization
library. It exports the Dump and Load functions for converting Perl data
structures to YAML strings, and the other way around.

YAML is a generic data serialization language that is optimized for
human readability. It can be used to express the data structures of most
modern programming languages.