Add a port for Data::Hierarchy. From DESCR:
Data::Hierarchy provides a simple interface for manipulating inheritable
data attached to a hierarchical environment (like a filesystem).

One use of Data::Hierarchy is to allow an application to annotate paths
in a real filesystem in a single compact data structure. However, the
hierarchy does not actually need to correspond to an actual filesystem.