Add a port for Class::Autouse. Needs the new Scalar-List-Utils. From DESCR:
"Class::Autouse" allows you to specify a class the will only load when a
method of that class is called. For large classes that might not be used
during the running of a program, such as Date::Manip, this can save you
large amounts of memory, and decrease the script load time a great deal.