Add a port for PathTools, another base module with a higher revision. From DESCR:
This package contains the Cwd and File::Spec modules.

Cwd provides functions for determining the pathname of the current
working directory.

File::Spec is designed to support operations commonly performed on file
specifications (usually called "file names", but not to be confused with
the contents of a file, or Perl's file handles), such as concatenating
several directory and file names into a single path, or determining
whether a path is rooted.