Add a port for Calendar::Simple. From DESCR:
A very simple module that exports one function called calendar.

This function returns a data structure representing the dates in a
month. The data structure returned is an array of array references. The
first level array represents the weeks in the month. The second level
array contains the actual days. By default, each week starts on a Sunday
and the value in the array is the date of that day. Any days at the
beginning of the first week or the end of the last week that are from
the previous or next month have the value undef.