Add a port for Class::Inspector. From DESCR:
Class::Inspector allows you to get information about a loaded class.
Most or all of this information can be found in other ways, but they
are not always very friendly, and usually involve a relatively high
level of Perl wizardry, or strange and unusual looking code.
Class::Inspector attempts to provide an easier, more friendly interface
to this information.