Add Subversion to CVS repository converter, requested by Leif Gruenwoldt in IRC (#cvs), written in mksh.
Pretty tested, pretty complete, optionally tags every SVN revision (so
that they can directly be checked out); preserves users and dates and
sets the RCS State to the SVN revision the RCS revision was changed at.

No branch handling (necessary) because SVN doesn’t do branches.

Use -B and -H options to have a restartable conversion.

The svn:executable (set or unset) and svn:mime-type (application/* or not)
properties for the LAST checked out revision are retained in the ,v files.

Please send improvements and/or requests to me!