Add a port for Sub::Uplevel. From DESCR:
Sub::Uplevel allows you to apparently run a function in a higher stack
frame.

It is like Tcl's uplevel() function, but not quite so dangerous. The
idea is just to fool caller(). All the really naughty bits of Tcl's
uplevel() are avoided.

THIS IS NOT THE SORT OF THING YOU WANT TO DO EVERYDAY.