Add a port for Array::Compare. From DESCR:
If you have two arrays and you want to know if they are the same or
different, then Array::Compare will be useful to you.

In addition to the simple comparison (which returns true
if the arrays are the same and false if they're different) there is also
a full comparison which returns a list containing the indexes of
elements which differ between the two arrays. If the arrays are the same
it returns an empty list. In scalar context the full comparison returns
the length of this list (i.e. the number of elements that differ).