Add a port for DateTime::TimeZone. Frm DESCR:
This class is the base class for all time zone objects. A time zone is
represented internally as a set of observances, each of which describes
the offset from GMT for a given time period.

Note that without the "DateTime.pm" module, this module does not do
much. It's primary interface is through a "DateTime" object, and most
users will not need to directly use "DateTime::TimeZone" methods.