Add a port for DateTime. From DESCR:
DateTime is a class for the representation of date/time combinations,
and is part of the Perl DateTime project. For details on this project
please see <http://datetime.perl.org/>. The DateTime site has a FAQ
which may help answer many "how do I do X?" questions. The FAQ is at
<http://datetime.perl.org/?FAQ>.

It represents the Gregorian calendar, extended backwards in time before
its creation (in 1582). This is sometimes known as the "proleptic
Gregorian calendar". In this calendar, the first day of the calendar
(the epoch), is the first day of year 1, which corresponds to the date
which was (incorrectly) believed to be the birth of Jesus Christ.

The calendar represented does have a year 0, and in that way differs
from how dates are often written using "BCE/CE" or "BC/AD".

For infinite datetimes, please see the DateTime::Infinite module.