New port: splix-2.0.0-0. From DESCR:
This driver is usable by all printer devices which understand the QPDL
(Quick Page Description Language) language. It covers several Samsung,
Xerox, Dell, and Ricoh printers like the ML-2250 which uses it as a
native printer language (Check the technical documentation for more
information).

If you are upgrading your current version, you MUST delete your printer
in the CUPS interface and install it again. This operation MUST be done
since CUPS make a copy of the PPD file and this file is modified between
versions.