Manually add calls to setpagedevice to set the page size, translate so that the bounding box starts at [ 0 0 ].
I think that calling setpagedevice is illegal in an EPS file but this is
what is needed for ps2pdf to create PDFs with a correct size. ps2pdf
disregards the bounding box size and uses a page size instead, so the
bounding box is now equal to the page size.