work around a weird problem:
On 2005-12-07 I made an ISO (500-some MiB) of the snapshot,
bootable and everything, which I distributed. I also burnt
the ISO to a CD-R medium (to more than one, but I gave the
others away, for free even).

I tested the ISO file before distributing, and everything
worked fine. I can also vnconfig(8) it and disklabel(8)
shows a correct slice table.

I installed from the medium I burnt on HERC, and no problem
occured, with an UltraPLEX 24x SCSI-CD ROM drive.

The very same physical medium in ODEM showed a weird beha-
viour which I forgot until bsiegert@ had the same while
installing from the ISO file onto MS Virtual PC for Mac:
# disklabel cd0
disklabel: can't read partition table: Invalid argument

The workaround in the installer is, that, when no disklabel
can be read and we know it's supposed to be a CD-ROM, to
just assume the 'c' slice is a correct ISO9660 filesystem.