Add a new (8 Kib) RSA key to be used for code signing later. We don't use the new key yet, just distribute it (as part of the base system in traditional ssh form, and as x.509 certi- ficate in ssl PEM form on the website, CAcert.org signed) in order to be able to switch to it later.
Tonnerre Lombard says that Ron Rivest says (or, was it Bruce
Schneier?) to use 6000-8000 bit lengths for new keys. Thus a
compromise, 8192 bit and 65537 as public exponent value were
used in generation of this key.

Key policy: it's being used on the CVS server (which is pub-
lically accessible) and tg@'s developer laptop. While I keep
the secret key file protected, I disclaim liability for sig-
natures not done by me (in case of doubt ask me for the RMD-
160 hash of the file in question, since gzsig(1) is not able
to tell you which file you have, only the file is ok).