Update firesomething to 1.5.0.3-0.
Our build is "unofficial" for trademark reasons and thus calls itself
"Deer Park" (the code name). The executable is called "firesomething", as
usual.

This probably needs -current, but it runs on my own -stable too (I have
rebuilt ld and src/lib from source). Don't even bother with it on #7.