Requested by bsiegert@ via eMail and <homsn> via IRC: During installation, ask the user whether he wants to send information about the hardware used to bsdstats.org and can disable that, in which case rc.conf.local will be amended with bsdstats_flags="" (empty), the default value is "YES". To entirely disable sending ANY information to bsdstats.org, the admin has to set it to "NO" in /etc/rc.conf.local. A deletion of /usr/share/misc/bsdstats has already resulted in disabling the code, and continues to do so.
bsiegert@ disagrees with this practice and wants the installation-time
question to be able to deactivate sending any information. I heard and
noticed the complaint and hereby acknowledge it, but decided to never-
theless go by with this method. I have discussed it through with Desk-
topBSD's Daniel Seuffert in Essen (on a conference), and Marc G. Four-
nier (the bsdstats.org man) has agreed to our proposal.

Please note that the information sent is *only*:
* one of your public IPv4 (or IPv6) addresses
* during the first run, 32 bytes of data from /dev/arandom used to ob-
  tain a "token"
* during all runs, the "key" from above and the "token" returned (this
  is a measure to estimate the number of MirOS installations)
* during all runs, the version of the operating system, i.e.
  - MirOS #9{semel,bis,ter,quater,...}
  - MirOS #9-stable
  - MirOS #9-current
  - MirOS #9-livecd
* unless deactivated, the output of "sysctl hw.model"
* unless deactivated, in the future (not yet specified) PCI device in-
  formation (most likely: PCIID (8 hexadecimal digits that are the sa-
  me for all exemplars/specimen of a type of hardware device), name of
  the device driver (e.g. vga for any vga(4) device), nothing else).
  This is, however, not yet implemented.

The MirOS Project does not have access to the raw data. We do not have
access to data the general public has access to. Only output of stati-
stics is ever published, no keys, tokens or IP addresses. See the web-
site of BSDstats for details. We try hard to not invade privacy.