Implement a new feature: * new sysctl machdep.apvreset (CPU_APVRESET) default value: 1 * new global: void (*panic_hook)(void) = NULL; <sys/systm.h> * new global: void (*panic_hook_display)(void) = NULL; (not declared) * panic(): if (panic_hook != NULL) call it before printf() * vga: on init, panic_hook_display = vga_panic_hook; * aperture: on open, IF machdep.apvreset != 0, panic_hook = panic_hook_display;   on close, panic_hook = NULL;
The vga_panic_hook() function at the moment does nothing (only
a debugging printf), but will contain code to reset the VGA video
mode to a sane, text mode, state. The code is of course only af-
fected if XF86 is active, and this is (as of yet) i386 specific.

This will help people dropping from X into ddb(4) and not seeing anything.
Not yet, though.