no functional change: introduce a (local) definition MIR18N_C_CSET that covers our "historical ISO-8859-1 kludge" for compatibility.
At the moment, this is set to 2, which basically means, the 'C'
locale is actually an 'en_US.ISO-8859-1' locale. If this is set
to 1, the isalpha() etc. functions and nl_langinfo(CODESET) will
return ASCII, but conversion functions, such as btowc(3), still
accept and/or return latin1 characters. If set to 0, 7-bit ASCII
will be enforced for the 'C' locale, as per its specification.

Because we only have two locales, 'C' and 'en_US.UTF-8', this
means that we _will_ go the Plan 9 way of eventually switching
totally to UTF-8. I think after wscons has been changed to some
kind of uwscons, we can make the switch to 0, but for now, I'll
leave it at 2.

Note:	for future MirOS-current snapshots, this may be lowered
	to 1, and raised back to 2 for releases, until suffici-
	ently tested and ports are converted.