Instead of defaulting | PKGNAME?=		${DISTNAME}-0 default it to | DASH_VER?=		0 | PKGNAME?=		${DISTNAME}-${DASH_VER}
I chose to take the name DASH_VER here because it's been
in wider use already (albeit in a non-standardised manner)
in MirPorts, and I've seen it elsewhere, I think. In my
own ports I used to use PL but that's too short and not
explicit enough for a new global variable.