* don't pass CPPFLAGS (as part of CFLAGS) to Build.sh
Benny's problem - described in the attachment of
 Message-ID: <BB26E796-4E24-4EFD-8A0B-CE4C58952D60@gmail.com> -
basically was caused by (going backwards):
* cond.c requires MAGICSLOTS2 and friends to be defined
* these would be found in condhashconsts.h
* which was generated by "generate"
* that required libmirmake to be linked in
* which, in turn, had not been built
Up to this point, commitid 1004593FFC84B90031B fixed the fact
that the build process wasn't aborted after a failure to build
libmirmake. (I wonder why I explicitly allowed the build to
continue without a libmirmake earlier. Anyway, scheduled for
rewrite, it is.) But libmirmake had not been built because...
| arc4random.c:43: error: conflicting types for 'arc4random_pushb'
| arc4random.c:38: error: previous implicit declaration of
|	'arc4random_pushb' was here
Now, why is there an "implicit declaration" - after all, these
are defined in contrib/code/mirmake/dist/contrib/mirmake.h!
Guessing from the ${CPPFLAGS}... the _already installed_ exemplar
of mirmake.h was included BEFORE ours° and did not yet contain
these prototypes.

The _correct_ solution is for MirPorts to _not_ pass the old
*part of* CPPFLAGS which says "-include ${.SYSMK:Q}/mirmake.h"
to Build.sh - but it says, depending on which version of MirMake
is already installed, one of:
* -isystem ${.SYSMK:Q} -include ${.SYSMK:Q}/mirmake.h
* -isystem "${.SYSMK}" -include "${.SYSMK}/mirmake.h"
* -isystem "@@@@shmk@@@@" -include "@@@@shmk@@@@/mirmake.h"
* -include "@@@@shmk@@@@/mirmake.h"
* -imacros "@@@@shmk@@@@/mirmake.h"
so we're currently removing all of CPPFLAGS, which _may_ be
more in MirPorts, but for the devel/mirmake port, is just the
above plus a "-idirafter ${LOCALBASE}/include", which we can
ignore because MirMake is self-hosted in that regard... for
now. (I can assume additional prerequisites on Interix - but
on the other hand, they'd probably be better placed in the
libmirmake itself then, which is kind of a libbsdlibc, or a
libshouldbeinlibc, or whatever.)

Now, I can't just use ${CFLAGS:S/${CPPFLAGS}//} because make(1)
does word splitting on variables... but if I first quote (:Q)
CFLAGS, I get one word, and can sed out a quoted CPPFLAGS. This
is also some source of suckage.

°) The final solution which will be implemented later is,
to use constructs like
| OUR_CPPFLAGS=
| OUR_CPPFLAGS="$OUR_CPPFLAGS -Dfoo"
| OUR_CPPFLAGS="$OUR_CPPFLAGS -Dbar"
| CPPFLAGS="$OUR_CPPFLAGS $CPPFLAGS"
instead of the current
| CPPFLAGS="$CPPFLAGS -Dfoo"
| CPPFLAGS="$CPPFLAGS -Dbar"
in the top half of Build.sh, thusly placing ours in front of,
not after, the CPPFLAGS inherited from the build environment.
(I should also place some of them, e.g. Interix' -D_ALL_SOURCE,
in the final ${CPPFLAGS} in <*.mk> if we're going to provice a
kind-of-BSD environment... or, if we're staying "lean", leave
things like that to MirPorts' own <*.mk> files.)