change the code to be not only standards-conforming, but have additional functions guarantees: * you can mix wide-oriented and byte-oriented I/O functions   (XXX really?) * if fgetwc(3) and friends return WEOF, set the error indicator   and set errno to EILSEQ, input processing can continue. The   first byte of the invalid multi-byte sequence is eaten in any   case; if there were more than one octet processed, the last   one is pushed back so that it can be re-retrieved and at most   one multibyte character is lost
the standards say these two are undefined, so why shouldn't we
be so liberal and define them to something that actually makes
sense? (I could even implement an unget buffer so that all but
the first octet are eaten, but that'd slow down I guess.) Code
in the base system is hereby allowed to use this, and all fol-
lowing, enhancements to the standards, because MirOS is a core
oriented "comes as a package" OS and we do not need to require
our own tools to only use normed functions unless it's made to
a portable package.