if we encounter a three-byte sequence, where the first two are valid 3-byte multibyte sequence beginning and middle char and the last one is not a valid 3-byte multibyte sequence end char, do not only ungetc(3) the last but also the middle byte, which is won back from the mbstate_t information.
This works because:
* the encoding is UCS-2 (wide char) and UTF-8 (multibyte)
* mbstate_t encoding is like we use it
* non-minimalistic encoding errors are handled as EILSEQ
* EILSEQ returns without changing mbstate_t *ps argument

Effect: <valid> <valid> <invalid> now yields WEOF WEOF <…>
instead of WEOF <…> (where <…> is what the invalid char
starts), therefore, when WEOF+EILSEQ is cought and mapped
into U+FFFD and retry, giving a more original-like idea
of how the input looked like.

This rollback could be adapted for UCS-4 + UTF-8, but
that's outside of my scope.