• generate a shared libossaudio
This sucks, but libtool in pkgsrc® can't build a shared libesd if it can't
link against this (static deplib, stuff like that), and if I just install
the piclib, which is supposed to work, it links against the non-pic static
archive, using -lossaudio though. Correct approach to fix this is to have
the {compiler driver,linker} look for libFOO_pic.a on -lFOO _only_ if ge-
nerating a shared library (I think). Or am I mistaken?