nuke the sparc “miniroot”, installboot(8/sparc) didn’t really like it, and I don’t even exactly know what it is
we’ll make sparc have a floppy ramdisk kernel and a cdrom/bsd.rd.net/bsd.rd
kernel like i386, except that we can’t use a GENERIC for the latter
for now we use the former, which is 300K or so too big for a floppy though,
as the latter and won’t produce a floppy

I want #10 to contain sparc, not in a #10bis to be produced later. That bis
version is to fix these bugs. I want sparc to be make work, and convert all
Debian/sparc32 users to MirBSD.