Another step towards recomming arandom(4) for high-quality high-security use (and arc4random(3) gets improved as well):
As suggested by Bruce Schneier’s Yarrow paper, split the arandom(4) source
into a “slow pool” and a “fast pool”, the latter being what we already had,
the former being the new thing: the uint8_t initial_entropy[16] array in-
troduced with commitid 10047E40AFE44E65982 is employed, after the initial
enqueue of its content into the pool and bzero(3)ing it (and setting the
value of initial_entropy_ptr from the initial -1 to 0), to keep a 128-bit
“slow pool” (with initial_entropy_ptr ranging from 0‥15) which is XOR’d
into the 256 octets stirring the arcfour state IFF it’s full. The “slow
pool” is filled whenever any entropy is dequeued from the main pool to an
amount that would deplete it, or on arc4 stirs if we’ve got an octet left
after eating up some pool data. The slow pool mixes into all of the fast
pool upon emptying.

While care has been taken on designing this, srandom(4) may be affected as
its check for the data size read being smaller than entropy_count comes be-
fore the dequeue action (which, in turn, _can_ extract eight more bits if
the pool would be emptied). Suggestion is to use very small reads, if you
must use srandom(4), or to use arandom(4) instead (or just mix both).