SECURITY: Update fetchmail to 6.3.8, with two security patches.
The fetchmail developers chose to distribute the patches in the WORST
way possible: embedding it, along with the description, in a PGP clearsigned
text file, which mangles "-" lines to "- -". Then they tell you to remove
these by hand. Oh, and the patches do not apply without -l (because of
mangled whitespace). Of course, one of these needs -p0 and the other one
-p1. Sigh.