document that, except for the first 512 bytes, stage2 and stage2_eltorito do not differ (they just use INT13,4B01 to ask… something we could do as well later) and as such, should work too
I have bad luck though, neither with -cdrom nor with -hda does it work for
me, but because of this I think it just doesn’t do ISO 9660 for some rea-
son or the other. (Maybe it tries to parse the Sun disklabel? Nah, that
isn’t it either. WTF.)