gecko2@ prodded me to investigate into a problem I had once (and no more afterwards) but could reproduce after he did: by trying to boot a fresh- ly installed from the FOSDEM 2009 ISO system.
The first stage boot loader (bootxx) would load 4 filesystem blocks of
16 KiB each == 128 sectors of the second stage boot loader (ldbsd.com),
because that’s what installboot(8) tells it – size in filesystem blocks.
Now the first stage boot loader is located at 4000:FE00h, whereas the
second stage boot loader begins at 4000:0000h. In the end, the last of
these 128 sectors ends up overwriting the code it loads it.

It did not appear for me any more because, due to sheer luck, the last
sector contained only no-ops (possibly 00h bytes), so it basically fell
through from 4000:FFFFh to 4000:0000h due to a 16 bit wrap.

This fix checks the bounds of the load address to see if the end of the
next sector load would wrap the boundary. Unless the size of a sector
were smaller than the size of bootxx, it would catch all possible cases.