The ability to generate a partition table and a partition for the grml manifold-boot™ ISO without having to sacrifice these 64 bytes of block table beforehand (bxinst only) by means of two new options: • -g C:H:S works like bxinst.sparc and initiates creating a partition • -M no (no=1‥4) makes room for the partition table
If neither -g nor -M are given, nothing is changed.
If -g and -M are given and valid, a partition is created with these values.
If -g is given but -M isn’t, it defaults to the last partition.
If -M is given but -g isn’t, the partition table is merely NUL’d.

The partition type passed from the -p option will be used.