tentative major bump to libm.so.7.0 which probably should have been done instead of a minor bump to 6.2 back then… but what few did we know… anyway, basically remove the “long double” remnants, since if you were using it on MirBSD, you were Doing It Wrong already anyway as sparc only has 32/64 bit float, and while i387 has 32/64/80 bit, our FPUCW limits precision to 64 bit (and the ABIs are at issue for whether “long double” should be stored with 80 or 96 bit and *THEN* SSE wants 128-bit alignment for them, so…)
This removes “long double” functions nobody should have been using,
on i386, which doesn’t really warrant a major bump, and __infinity,
which our new libm provides as GCC builtin (and __infinityf was not
included anyway).

BUT! Do not use this yet. This needs verification on sparc, where a
Perl build failed due to having *more* (but actually even uselesser
– heh) long double stubs…